Minor casting fix

+++ plugins/nrepellent-adaptive.c
@@ -89,7 +89,7 @@ static LV2_Handle instantiate_adaptive(const LV2_Descriptor *descriptor,
     return NULL;
   }
 
-  self->soft_bypass = signal_crossfade_initialize(self->sample_rate);
+  self->soft_bypass = signal_crossfade_initialize((uint32_t)self->sample_rate);
 
   if (!self->soft_bypass) {
     nrepel_free(self);
