+++ plugins/nrepellent.c
@@ -202,7 +202,7 @@ static LV2_Handle instantiate(const LV2_Descriptor *descriptor,
 
   self->sample_rate = (float)rate;
 
-  self->soft_bypass = signal_crossfade_initialize(self->sample_rate);
+  self->soft_bypass = signal_crossfade_initialize((uint32_t)self->sample_rate);
 
   if (!self->soft_bypass) {
     nrepel_free(self);
